ref: move basetask to nodebase
